requirements.txt
#Requirements PICC9

Flask
pillow
face-recognition
numpy
Pyrebase4
opencv_python
cvzone
firebase_admin
easyocr
rapidfuzz